+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -82,6 +82,7 @@ public let masterRuleList = RuleList(rules:
     TypeBodyLengthRule.self,
     TypeNameRule.self,
     ValidDocsRule.self,
+    ValidIBInspectableRule.self,
     VariableNameRule.self,
     VerticalWhitespaceRule.self
 )